make segmentation warning work for skeleton and volume tracing; refactor low seg warning for volume (#1912)
