+++ app/assets/javascripts/oxalis/model/sagas/skeletontracing_saga.js
@@ -40,7 +40,6 @@ import type {
 } from "oxalis/store";
 import type { UpdateAction } from "oxalis/model/sagas/update_actions";
 import api from "oxalis/api/internal_api";
-import Model from "oxalis/model";
 
 function* centerActiveNode() {
   getActiveNode(yield select(state => state.tracing)).map(activeNode => {
@@ -104,39 +103,7 @@ export function* watchSkeletonTracingAsync(): Generator<*, *, *> {
     ["SET_ACTIVE_TREE", "SET_ACTIVE_NODE", "DELETE_NODE", "DELETE_BRANCHPOINT", "SELECT_NEXT_TREE"],
     centerActiveNode,
   );
-  yield [watchBranchPointDeletion(), warnAboutSegmentationOpacity()];
-}
-
-function* warnAboutSegmentationOpacity(): Generator<*, *, *> {
-  let zoomStepWarningToast;
-  const warnMaybe = function() {
-    const shouldWarn = Model.shouldDisplaySegmentationData() && !Model.canDisplaySegmentationData();
-    if (shouldWarn && zoomStepWarningToast == null) {
-      const toastType = Store.getState().tracing.type === "volume" ? "danger" : "info";
-      zoomStepWarningToast = Toast.message(
-        toastType,
-        "Segmentation data and volume tracing is only fully supported at a smaller zoom level.",
-        true,
-      );
-    } else if (!shouldWarn && zoomStepWarningToast != null) {
-      zoomStepWarningToast.remove();
-      zoomStepWarningToast = null;
-    }
-  };
-
-  yield take("INITIALIZE_SETTINGS");
-  warnMaybe();
-
-  while (true) {
-    yield take([
-      "ZOOM_IN",
-      "ZOOM_OUT",
-      "ZOOM_BY_DELTA",
-      action =>
-        action.type === "UPDATE_DATASET_SETTING" && action.propertyName === "segmentationOpacity",
-    ]);
-    warnMaybe();
-  }
+  yield [watchBranchPointDeletion()];
 }
 
 function* diffNodes(
